commands: rename baseClasses and remove abstract
