+++ BlasII.CheatConsole/Properties/AssemblyInfo.cs
@@ -2,4 +2,4 @@ using BlasII.CheatConsole;
 using MelonLoader;
 
 [assembly: MelonInfo(typeof(BlasII.CheatConsole.Main), ModInfo.MOD_NAME, ModInfo.MOD_VERSION, ModInfo.MOD_AUTHOR)]
-[assembly: MelonOptionalDependencies("Modding API", "1.0.0", "UI Framework", "0.1.0")]
+[assembly: MelonOptionalDependencies("Modding API", "1.1.0", "UI Framework", "0.1.0")]
